pyproject.toml
[tool.poetry]
name = "gptpdf"
version = "0.0.15"
description = "Using GPT to parse PDF"
authors = ["Chen Li <lichenarthurdata@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"
repository = "https://github.com/CosmosShadow/gptpdf"
packages = [
    { include = "gptpdf" },
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
GeneralAgent = "^0.3.21"
shapely = "^2.0.1"
pymupdf = "^1.24.7"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"


[[tool.poetry.source]]
name = "PyPI"
priority="primary"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"